fix(build): build the bundle on install
When migrating an app to use vite, a syntax error in a source map file
in `dist` was causing builds to fail. Fixing how the bundle was built
addresses the issue.

To avoid encountering this issue in the future, I've opted to add a
`prepare` script that makes sure the bundle is built whenever the
package is installed.

As a result of addressing code rot when figuring out how to build the
bundle, I've updated dependencies and specified minimum node & npm
versions.